Automated GACOS Data Request Script
Overview;
This Python script automates the submission of requests to GACOS for atmospheric corrections in InSAR processing. GACOS limits each request to a maximum of 20 images, making manual submissions time-consuming when handling a large number of interferograms. This script helps streamline the process, especially for long-term deformation monitoring.


Current Status;
🔹 The script currently requires an update to automatically trigger the Submit button.
🔹 It uses Firefox for web interactions.
🔹 The script imports selenium and time libraries.


Usage;
The script requires a JSON input file with the format in ''config.json'' file. It should be in same directory which code is copied. The ''output_format'' parameter allows selecting between Binary grid and GeoTIFF. This script currently submits requests in ''Binary grid'' format. You can change it to ''Geotiff''.


Planned Improvements;
✅ Implement automatic submission without requiring manual interaction.
✅ Enhance error handling and logging for smoother operation.


Contributions & Feedback;
Contributions and feedback are welcome! Feel free to submit issues or pull requests to improve the script.